DirMgr: Detect that directory state is ready if it begins with all MDs
Previously, we'd only call PendingNetDir::upgrade_if_necesessary
when adding a microdescriptor.  But if it began already having all
of its descriptors (because we found them in the cache), we wouldn't
actually upgrade it to a PendingNetDir::Yielding, which would make
it unusable, and would make us schedule its reset time too far
in the future.

Fixes #802.